requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
